Add db

+++ TodoMAUI/Models/TodoModel.cs
@@ -9,7 +9,7 @@ namespace TodoMAUI.Models
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public byte Priority { get; set; }
+        public int Priority { get; set; }
         public bool IsDone { get; set; }
     }
 }
